+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Fixed
 
 ### Changed
+- Added hasThumbnail and disabled properties to the Tile component. 
+- Added as hover image preview icon for the Tile component.
 
 ### Breaking changes